Fix doc build
