+++ doc/index.rst
@@ -19,13 +19,8 @@ cdflib requires python 3 and numpy. To install run
    :caption: Contents:
 
    introduction
-   modules/CDF function
-   modules/cdfread
-   modules/cdfwrite
-   modules/epochs
-   modules/epochs_astropy
-   modules/cdf_to_xarray
-   modules/xarray_to_cdf
+   modules/cdflib
+   modules/xarray
    development
