--- doc/modules/CDF function.rst
@@ -1,7 +1,0 @@
-CDF
-=====
-
-CDF function
---------------
-
-.. autofunction:: cdflib.CDF
