--- doc/modules/xarray_to_cdf.rst
@@ -1,11 +1,0 @@
-xarray_to_cdf
-=============
-
-xarray_to_cdf function
------------------------
-
-This module is comprised of a single function which takes XArray Dataset files and converts them into CDF files.  
-
-This will attempt to conform to `ISTP Compliance <https://spdf.gsfc.nasa.gov/istp_guide/istp_guide.html>`_, and optionally provide warnings if it is not compliant.  
-
-.. autofunction:: cdflib.xarray_to_cdf
